WORDS-53 todo: This test required finishing
